+++ src/components/pages/campaigns-details/components/batches-table/types.tsx
@@ -5,6 +5,7 @@ export type TProps = {
   campaignId: string
   title: string
   sdk: boolean
-  downloadLinks: (batch_id: string | number, campaign_id: string, title: string, encryptionKey?: string) => void
-  encryptionKey: string
+  downloadLinks: (batch_id: string | number, campaign_id: string, title: string, tokenAddress: string | null, encryptionKey?: string) => void
+  encryptionKey: string,
+  tokenAddress: string | null
 }