+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added info screen if network is not supported
 - Added Alchemy API to upload NFT tokens list
 - Added select component to choose available token ids for ERC721/ERC1155 campaign
+- Dynamic definition of user comission with contract instance
 
 ## 2.1.4
 - Restrict networks with env variable